update where and when logs are written
